fix: moved ServiceStatus to a new package, adjusted one of the methods to track where server shutdown was called from
