+++ server/src/main/java/com/hedera/block/server/BlockNodeAppInjectionModule.java
@@ -18,6 +18,8 @@ package com.hedera.block.server;
 
 import com.hedera.block.server.config.BlockNodeContext;
 import com.hedera.block.server.metrics.MetricsService;
+import com.hedera.block.server.service.ServiceStatus;
+import com.hedera.block.server.service.ServiceStatusImpl;
 import com.swirlds.config.api.Configuration;
 import dagger.Binds;
 import dagger.Module;
