+++ server/src/main/java/module-info.java
@@ -15,6 +15,7 @@ module com.hedera.block.server {
     exports com.hedera.block.server.health;
     exports com.hedera.block.server.persistence;
     exports com.hedera.block.server.notifier;
+    exports com.hedera.block.server.service;
 
     requires com.hedera.block.stream;
     requires com.google.protobuf;
