+++ server/src/test/java/com/hedera/block/server/health/HealthServiceTest.java
@@ -19,7 +19,7 @@ package com.hedera.block.server.health;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.*;
 
-import com.hedera.block.server.ServiceStatus;
+import com.hedera.block.server.service.ServiceStatus;
 import io.helidon.webserver.http.HttpRules;
 import io.helidon.webserver.http.ServerRequest;
 import io.helidon.webserver.http.ServerResponse;
