+++ server/src/test/java/com/hedera/block/server/mediator/MediatorInjectionModuleTest.java
@@ -18,8 +18,8 @@ package com.hedera.block.server.mediator;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import com.hedera.block.server.ServiceStatus;
 import com.hedera.block.server.config.BlockNodeContext;
+import com.hedera.block.server.service.ServiceStatus;
 import com.hedera.block.server.util.TestConfigUtil;
 import com.hedera.hapi.block.SubscribeStreamResponse;
 import com.hedera.hapi.block.stream.BlockItem;
